backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.6
alembic>=1.12.1
pydantic[email]>=2.9.0
